+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.4.0] - 2022-01-15
 ### Added
 - Add property `not` same result with method `not()`.
 - Add validation (rule) condition `where()`, only execute rule validation if condition as true (left side validation).